requirements.txt
retry==0.9.2
requests==2.25.1
requests-oauthlib==1.3.0	
tweepy==3.10.0
PyYAML~=5.4.1